Add new components for different condition types.
